requirements.txt
asgiref==3.2.10
boto3==1.14.36
botocore==1.17.36
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Django==2.2.15
docutils==0.15.2
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
google-api-core==1.22.0
google-auth==1.20.1
google-cloud-speech==1.3.2
googleapis-common-protos==1.52.0
googletrans-temp==2.3.0
grpcio==1.31.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
jdcal==1.4.1
jmespath==0.10.0
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.0.3
moviepy==1.0.3
numpy==1.18.5
openpyxl==2.6.4
packaging==20.4
pandas==0.24.2
Pillow==7.2.0
proglog==0.1.9
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scipy==1.4.1
seaborn==0.9.1
sentencepiece==0.1.91
six==1.15.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
sqlparse==0.3.1
tokenizers==0.8.1rc1
torch==1.5.1
tqdm==4.48.2
transformers==2.5.1
urllib3==1.25.10
